Fixed issue that was causing build failure
